# docs/swagger.yaml
definitions:
...
  types.PodListResponse:
    properties:
      age:
        type: string
      clusterName:
        type: string
      configName:
        type: string
      cpu:
        type: string
      hasUpdated:
        type: boolean
      lastRestartAt:
        type: string
      lastRestartReason:
        type: string
      memory:
        type: string
      name:
        type: string
      namespace:
        type: string
      node:
        type: string
      podIP:
        type: string
      qos:
        type: string
      ready:
        type: string
      restarts:
        type: string
      status:
        type: string
      uid:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/pod/{name}/exec/ws:
    get:
      consumes:
      - application/json
      description: Execute interactive commands in a pod container by name via WebSocket
        connection
      parameters:
      - description: Pod name
        in: path
        name: name
        required: true
        type: string
      - description: Namespace name
        in: query
        name: namespace
        required: true
        type: string
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Container name (defaults to first container)
        in: query
        name: container
        type: string
      - description: 'Command to execute (default: /bin/sh)'
        in: query
        name: command
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket connection established
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Execute Commands in Pod by Name via WebSocket
      tags:
      - WebSocket
...
  /api/v1/pods:
    get:
      consumes:
      - application/json
      description: Get all pods with real-time updates via Server-Sent Events
      parameters:
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Namespace filter
        in: query
        name: namespace
        type: string
      - description: Node name filter
        in: query
        name: node
        type: string
      - description: Owner type (deployment, daemonset, etc.)
        in: query
        name: owner
        type: string
      - description: Owner name
        in: query
        name: ownerName
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: Stream of pod data
          schema:
            items:
              $ref: '#/definitions/types.PodListResponse'
            type: array
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KubeConfig: []
      summary: Get Pods (SSE)
      tags:
      - Workloads
  /api/v1/pods/{namespace}/{name}/exec/ws:
    get:
      consumes:
      - application/json
      description: Execute interactive commands in a pod container via WebSocket connection
      parameters:
      - description: Namespace name
        in: path
        name: namespace
        required: true
        type: string
      - description: Pod name
        in: path
        name: name
        required: true
        type: string
      - description: Kubernetes configuration ID
        in: query
        name: config
        required: true
        type: string
      - description: Cluster name
        in: query
        name: cluster
        type: string
      - description: Container name (defaults to first container)
        in: query
        name: container
        type: string
      - description: 'Command to execute (default: /bin/sh)'
        in: query
        name: command
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: WebSocket connection established
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - KubeConfig: []
      summary: Execute Commands in Pod via WebSocket
      tags:
      - WebSocket
...
swagger: "2.0"